refactor: implement mutation rule handlers (#6684)
* refactor: implement mutation rule handlers


* fix


* fix


---------
