+++ pkg/engine/handlers/handler.go
@@ -6,6 +6,7 @@ import (
 	"github.com/go-logr/logr"
 	kyvernov1 "github.com/kyverno/kyverno/api/kyverno/v1"
 	engineapi "github.com/kyverno/kyverno/pkg/engine/api"
+	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 )
 
 type Handler interface {
@@ -13,6 +14,21 @@ type Handler interface {
 		context.Context,
 		logr.Logger,
 		engineapi.PolicyContext,
+		unstructured.Unstructured,
 		kyvernov1.Rule,
-	) *engineapi.RuleResponse
+		func(logr.Logger, engineapi.PolicyContext, kyvernov1.Rule) *engineapi.RuleResponse,
+	) (unstructured.Unstructured, []engineapi.RuleResponse)
+}
+
+func RuleResponses(rrs ...*engineapi.RuleResponse) []engineapi.RuleResponse {
+	var out []engineapi.RuleResponse
+	for _, rr := range rrs {
+		if rr != nil {
+			out = append(out, *rr)
+		}
+	}
+	if len(out) == 0 {
+		return nil
+	}
+	return out
 }
